// src/typings/api/assets/assets.d.ts
export type AssetsRecordsListReq = {
  /** 类型，201. 充币，202. 提币 ,203 pay , 204. 冲正，501、现货手续费 502、合约手续费 503、提币手续费 504、锁仓手续费 */
  type?: number[] | string;
  /** 开始时间 */
  startDate?: number;
  /** 结束时间 */
  endDate?: number;
  /** 币种 id */
  coinId?: string;
  /** 状态，1、进行中 2、成功 3、失败 4、错误 */
  status?: number[];
  /** 页数 */
  pageNum: number;
  /** 条数 */
  pageSize: number;
  /** 财务类型：logTypeCd：全部时不传，1，现货，2、充提、3、借还款、4、合约、5、手续费、6、衍生品、7、其他 */
  logType?: number;
  /** 合约组 id */
  groupId?: string;
  /** 代理商返佣类型 */
  rebateType?: string[];
}

export type AssetsRecordsListResp = {
  pageNum: number;
  pageSize: number;
  total: number;
  list: RecordsListResp[];
}

export type RecordsListResp = {
  id: string;
  /** 币种 */
  businessCoin: string;
  /** 总金额 */
  total: string;
  /** 创建时间 */
  createdByTime: number;
  /** 状态，1、进行中 2、成功 3、失败 4、错误 */
  status: number;
  /** 类型，201. 充币，202. 提币 ,203 pay , 204. 冲正，501、现货手续费 502、合约手续费 503、提币手续费 504、锁仓手续费 */
  type: number;
  /** 合约类型 */
  groupType?: number;
  /** 手动/自动 */
  operationType?: string;
  /** 合约组名称 */
  groupName?: string;
  /** 划转来源 */
  from?: string;
  /** 划转去向 */
  to?: string;
  /** 订单 id */
  /** 返佣类型 */
  rebateTypeCd?: string;
  /** 现货订单 ID，现货买入卖出查询明细用 */
  orderId?: string;
}